pyproject.toml
[tool.poetry]
name = "scraping-test"
version = "0.1.0"
description = ""
authors = ["Alfredo Delgado <alfredodlgdmrn@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
bs4 = "^0.0.2"
pytest-playwright = "^0.5.2"
python-dotenv = "^1.0.1"
lxml = "^5.3.0"
pandas = "^2.2.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
